pyproject.toml
[tool.poetry]
name = "nb-alpha"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
lorem = "^0.1.1"
aiohttp = "^3.7.4"
cryptography = "^3.4.6"
dblpy = "^0.4.0"
motor = "^2.3.1"
praw = "^7.2.0"
pymongo = "^3.11.3"
flask = "^1.1.2"
bs4 = "^0.0.1"
tornado = "^6.1"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
